Убирает отслеживание фильтра в шапке, удаляет несохраненные изменения для точки маршрута
