+++ src/api/destinations-api-service.js
@@ -5,5 +5,5 @@ export default class DestinationsApiService extends ApiService {
   get destinations() {
     return this._load({ url: 'destinations' }).then(ApiService.parseResponse);
   }
-  
+
 }
