+++ src/main.js
@@ -18,6 +18,7 @@ import FilterModel from './model/filter-model.js';
 import NewTripPointButtonView from './view/new-trip-point-button-view.js';
 import FailedView from './view/failed-view.js';
 
+
 const AUTHORIZATION = 'Basic ls2itl0iot0acn2hd';
 const END_POINT = 'https://24.objects.htmlacademy.pro/big-trip';
