+++ src/presenter/filter-presenter.js
@@ -1,7 +1,7 @@
+import { FilterType, UpdateType } from '../const.js';
+import { filter } from '../utils/filter.js';
 import { render, replace, remove } from '../framework/render.js';
 import FilterView from '../view/filter-view.js';
-import { filter } from '../utils/filter.js';
-import { FilterType, UpdateType } from '../const.js';
 
 
 export default class FilterPresenter {
